chore: minor tweaks on Rust FFI interface
Prepare the Rust FFI code for better scaling.
