+++ src/lean/sarray.rs
@@ -1,18 +1,18 @@
-use std::{
-    alloc::{alloc, handle_alloc_error, Layout},
-    ptr,
-};
+use std::alloc::Layout;
 
-use super::{Array, LeanObject};
+use super::{alloc_layout, object::LeanObject, CArray};
 
+/// ```c
+/// #define LeanScalarArray 248
+/// ```
 const LEAN_SCALAR_ARRAY: u8 = 248;
 
 /// ```c
 /// typedef struct {
-/// lean_object   m_header;
-/// size_t        m_size;
-/// size_t        m_capacity;
-/// uint8_t       m_data[];
+///     lean_object m_header;
+///     size_t      m_size;
+///     size_t      m_capacity;
+///     uint8_t     m_data[];
 /// } lean_sarray_object;
 /// ```
 #[repr(C)]
@@ -20,7 +20,7 @@ pub struct LeanSArrayObject {
     pub m_header: LeanObject,
     pub m_size: usize,
     pub m_capacity: usize,
-    pub m_data: Array<u8>,
+    pub m_data: CArray<u8>,
 }
 
 impl LeanSArrayObject {
@@ -37,17 +37,15 @@ impl LeanSArrayObject {
         )
         .expect("Couldn't compute the memory layout");
 
-        let ptr = unsafe { alloc(layout) as *mut LeanSArrayObject };
-        if ptr.is_null() {
-            handle_alloc_error(layout);
-        }
+        let ptr = alloc_layout::<LeanSArrayObject>(layout);
 
-        let obj = LeanObject::new(1, 0, 1, LEAN_SCALAR_ARRAY);
+        let size_of_u8 = size_of::<u8>() as u8;
+        let header = LeanObject::new(1, 0, size_of_u8, LEAN_SCALAR_ARRAY);
 
         unsafe {
-            ptr::write(&mut (*ptr).m_header, obj);
-            ptr::write(&mut (*ptr).m_size, len);
-            ptr::write(&mut (*ptr).m_capacity, len);
+            (*ptr).m_header = header;
+            (*ptr).m_size = len;
+            (*ptr).m_capacity = len;
             (*ptr).m_data.copy_from_slice(slice);
         }
